requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
motor
httpx             # for async Telegram API calls
pydantic
beanie